requirements.txt
discord.py[voice]==1.7.3
python-dotenv==0.20.0